# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/email/send-invoice": {
      "post": {
        "operationId": "EmailController_sendInvoiceEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "invoiceId",
                  "toEmail"
                ],
                "properties": {
                  "invoiceId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  },
                  "toEmail": {
                    "type": "string",
                    "format": "email",
                    "example": "client@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Send invoice email with PDF attachment",
        "tags": [
          "Email"
        ]
      }
    },
    "/email/send-reminder": {
      "post": {
        "operationId": "EmailController_sendReminderEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "invoiceId",
                  "toEmail"
                ],
                "properties": {
                  "invoiceId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  },
                  "toEmail": {
                    "type": "string",
                    "format": "email",
                    "example": "client@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Send payment reminder email",
        "tags": [
          "Email"
        ]
      }
    },
    "/email/send-receipt": {
      "post": {
        "operationId": "EmailController_sendReceiptEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "paymentId",
                  "toEmail"
                ],
                "properties": {
                  "paymentId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  },
                  "toEmail": {
                    "type": "string",
                    "format": "email",
                    "example": "client@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Send payment receipt email",
        "tags": [
          "Email"
        ]
      }
    },
...
    "/ca/invite": {
      "post": {
        "operationId": "CaPortalController_inviteCA",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "permissions"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "ca@example.com"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "read_journals"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Invite a CA to access organization data",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/accept-invite": {
      "post": {
        "operationId": "CaPortalController_acceptInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Invitation token from the CA invite email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Accept a CA invitation",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/revoke/{userId}": {
      "delete": {
        "operationId": "CaPortalController_revokeAccess",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Revoke CA access to organization",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/connected": {
      "get": {
        "operationId": "CaPortalController_getConnectedCAs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Get all CAs connected to this organization",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/clients": {
      "get": {
        "operationId": "CaPortalController_getClients",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all organizations this CA has access to",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/clients/{orgId}/summary": {
      "get": {
        "operationId": "CaPortalController_getClientSummary",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get client organization summary dashboard",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/clients/{orgId}/export": {
      "get": {
        "operationId": "CaPortalController_exportClientData",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fy",
            "required": true,
            "in": "query",
            "description": "Fiscal year, e.g. 2025-26",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Export annual data pack for a client",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/annotations": {
      "post": {
        "operationId": "CaPortalController_addAnnotation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orgId",
                  "entityType",
                  "entityId",
                  "comment"
                ],
                "properties": {
                  "orgId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  },
                  "entityType": {
                    "type": "string",
                    "example": "journal_entry"
                  },
                  "entityId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "11111111-2222-3333-4444-555555555555"
                  },
                  "comment": {
                    "type": "string",
                    "example": "Please reclassify under utilities."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Add a CA annotation on a financial entity",
        "tags": [
          "CA Portal"
        ]
      },
      "get": {
        "operationId": "CaPortalController_getAnnotations",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List CA annotations",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/approvals": {
      "get": {
        "operationId": "CaPortalController_getApprovalQueue",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get pending journal entries awaiting CA approval",
        "tags": [
          "CA Portal"
        ]
      }
    },
    "/ca/approvals/{journalEntryId}": {
      "post": {
        "operationId": "CaPortalController_approveEntry",
        "parameters": [
          {
            "name": "journalEntryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orgId",
                  "approved"
                ],
                "properties": {
                  "orgId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  },
                  "approved": {
                    "type": "boolean",
                    "example": true
                  },
                  "comment": {
                    "type": "string",
                    "example": "Approved after verification."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Approve or reject a journal entry",
        "tags": [
          "CA Portal"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kontafy API",
    "description": "REST API for Kontafy — cloud-native accounting for Indian businesses. All paths are versioned: prefix every request with `/v1`. Authenticated routes expect `Authorization: Bearer <access_token>` and most multi-tenant routes also expect `X-Org-Id: <organization id>`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.kontafy.com/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:5002/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /v1/auth/login"
      },
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id",
        "description": "Organization id for multi-tenant routes"
      }
    },
...
  }
}